Sort moved and name changes
Moved sorting to final print
Fixed naming in dijkstra both in code and name of c file/h file
